refactor(codec): extract field encode/decode emitters into CodecEmitterFields.kt (step 6)
Move the per-field-type encode/decode emit family out of CodecEmitter into a
new package-internal top-level file: every appendDecode* / appendEncode*
helper for the supported field shapes (scalar, value-class scalar,
conditional, length-prefixed string/message/UseCodec, @LengthFrom,
@RemainingBytes, bare @UseCodec / @ProtocolMessage scalars) plus the shared
scalarHeaderBytes / appendBufferPrefix* utilities.

The orchestration dispatchers (appendDecodeField / appendEncodeField and the
batching layer) stay in CodecEmitter and call these unqualified across the
package — the same cross-file pattern as steps 4-5. The family's one straggler
dependency, the stateless naturalScalarWriteStatement, moves alongside its read
sibling naturalScalarReadExpr in CodecEmitterScalarUtils.kt.

All moved verbatim (no batchCounter/codeGenerator/logger state); CodecEmitter
3862 -> 1911 lines, CodecEmitterFields 1951. Byte-identical codegen verified by
CodecSnapshotTest (jvmTest FROM-CACHE = identical KSP output); processor +
codec-test jvmTest suites green.
