+++ buffer-codec-processor/src/main/kotlin/com/ditchoom/buffer/codec/processor/CodecEmitter.kt
@@ -183,28 +183,6 @@ internal class CodecEmitter(
     }
 
     /**
-     * Write statement for a natural-width scalar given an
-     * accessor expression. Boolean encodes as `0x00` / `0x01`.
-     */
-    private fun naturalScalarWriteStatement(
-        kind: ScalarKind,
-        accessor: String,
-    ): String =
-        when (kind) {
-            ScalarKind.Boolean -> "buffer.writeByte(if ($accessor) 1.toByte() else 0.toByte())"
-            ScalarKind.UByte -> "buffer.writeUByte($accessor)"
-            ScalarKind.UShort -> "buffer.writeUShort($accessor)"
-            ScalarKind.UInt -> "buffer.writeUInt($accessor)"
-            ScalarKind.ULong -> "buffer.writeULong($accessor)"
-            ScalarKind.Byte -> "buffer.writeByte($accessor)"
-            ScalarKind.Short -> "buffer.writeShort($accessor)"
-            ScalarKind.Int -> "buffer.writeInt($accessor)"
-            ScalarKind.Long -> "buffer.writeLong($accessor)"
-            ScalarKind.Float -> "buffer.writeFloat($accessor)"
-            ScalarKind.Double -> "buffer.writeDouble($accessor)"
-        }
-
-    /**
      * /14c — `@FramedBy` file spec. Emits an `object`
      * codec with the new encode signature
      * (`encode(value, context, factory): ReadBuffer`) and the strict
@@ -1929,1934 +1907,5 @@ internal class CodecEmitter(
             }
         }
         return builder.build()
-    }
-
-    private fun scalarHeaderBytes(shape: CodecShape): Int = shape.fields.sumOfFixedWireBytes().requireFixed("scalarHeaderBytes")
-
-    private fun appendDecodeScalar(
-        body: CodeBlock.Builder,
-        field: FieldSpec.Scalar,
-    ) {
-        val widthMatches = field.wireBytes == field.kind.width
-        val explicitOrder = field.resolvedWireOrder != Endianness.Default
-        // Natural-width Default — trust buffer.byteOrder.
-        if (widthMatches && !explicitOrder) {
-            body.addStatement("val %L = %L", field.name, naturalScalarReadExpr(field.kind))
-            return
-        }
-        // Natural-width explicit Big/Little on a multi-byte scalar. Read at
-        // the natural width and canonicalize via swapBytes when buffer.byteOrder
-        // differs from the wire order. Matches the batched single-field code
-        // shape — single readShort/readInt/readLong instead of N readUByte +
-        // shift/or assembly. (1-byte scalars fall through to the manual path
-        // since they have no byte order; the manual path emits a single byte
-        // read for that case.)
-        if (widthMatches && explicitOrder && field.kind.width > 1) {
-            appendNaturalReadWithSwap(body, field)
-            return
-        }
-        val bigEndian =
-            when (field.resolvedWireOrder) {
-                Endianness.Little -> false
-                Endianness.Big, Endianness.Default -> true
-            }
-        appendManualScalarDecode(body, field, bigEndian)
-    }
-
-    private fun appendNaturalReadWithSwap(
-        body: CodeBlock.Builder,
-        field: FieldSpec.Scalar,
-    ) {
-        val canonicalOrder =
-            if (field.resolvedWireOrder == Endianness.Big) "BIG_ENDIAN" else "LITTLE_ENDIAN"
-        val readMethod =
-            when (field.kind.width) {
-                2 -> "readShort"
-                4 -> "readInt"
-                8 -> "readLong"
-                else -> error("unsupported natural width ${field.kind.width}")
-            }
-        val rawVar = "${field.name}Raw"
-        body.addStatement("val %L = buffer.%L()", rawVar, readMethod)
-        when (field.kind) {
-            ScalarKind.Short, ScalarKind.Int, ScalarKind.Long ->
-                body.addStatement(
-                    "val %L = if (buffer.byteOrder == %T.%L) %L else %M(%L)",
-                    field.name,
-                    BYTE_ORDER_CN,
-                    canonicalOrder,
-                    rawVar,
-                    SWAP_BYTES_MN,
-                    rawVar,
-                )
-            ScalarKind.UShort, ScalarKind.UInt, ScalarKind.ULong -> {
-                val toUnsigned =
-                    when (field.kind) {
-                        ScalarKind.UShort -> "toUShort"
-                        ScalarKind.UInt -> "toUInt"
-                        ScalarKind.ULong -> "toULong"
-                        else -> error("unreachable")
-                    }
-                body.addStatement(
-                    "val %L = (if (buffer.byteOrder == %T.%L) %L else %M(%L)).%L()",
-                    field.name,
-                    BYTE_ORDER_CN,
-                    canonicalOrder,
-                    rawVar,
-                    SWAP_BYTES_MN,
-                    rawVar,
-                    toUnsigned,
-                )
-            }
-            ScalarKind.Float ->
-                body.addStatement(
-                    "val %L = Float.fromBits(if (buffer.byteOrder == %T.%L) %L else %M(%L))",
-                    field.name,
-                    BYTE_ORDER_CN,
-                    canonicalOrder,
-                    rawVar,
-                    SWAP_BYTES_MN,
-                    rawVar,
-                )
-            ScalarKind.Double ->
-                body.addStatement(
-                    "val %L = Double.fromBits(if (buffer.byteOrder == %T.%L) %L else %M(%L))",
-                    field.name,
-                    BYTE_ORDER_CN,
-                    canonicalOrder,
-                    rawVar,
-                    SWAP_BYTES_MN,
-                    rawVar,
-                )
-            ScalarKind.UByte, ScalarKind.Byte, ScalarKind.Boolean ->
-                error("1-byte scalar should not take the natural-read-with-swap path")
-        }
-    }
-
-    private fun appendManualScalarDecode(
-        body: CodeBlock.Builder,
-        field: FieldSpec.Scalar,
-        bigEndian: Boolean,
-    ) {
-        val width = field.wireBytes
-        if (field.kind == ScalarKind.UByte && width == 1) {
-            body.addStatement("val %L = buffer.readUByte()", field.name)
-            return
-        }
-        if (field.kind == ScalarKind.Byte && width == 1) {
-            body.addStatement("val %L = buffer.readByte()", field.name)
-            return
-        }
-        // Assemble the wire bytes into a wide unsigned accumulator, then narrow
-        // to the field's declared kind. Signed kinds reinterpret the bit pattern
-        // via toShort()/toInt()/toLong() (Kotlin's UShort/UInt/ULong .toX() are
-        // bit-preserving). Float/Double go through fromBits().
-        val accumulator = if (width >= 5) "toULong" else "toUInt"
-        for (i in 0 until width) {
-            body.addStatement(
-                "val %L = buffer.readUByte().%L()",
-                "${field.name}B$i",
-                accumulator,
-            )
-        }
-        val parts =
-            (0 until width).map { i ->
-                val byteName = "${field.name}B$i"
-                val shiftBits = if (bigEndian) (width - 1 - i) * 8 else i * 8
-                if (shiftBits == 0) byteName else "($byteName shl $shiftBits)"
-            }
-        val combined = if (parts.size == 1) parts[0] else "(${parts.joinToString(" or ")})"
-        when (field.kind) {
-            ScalarKind.UByte -> body.addStatement("val %L = %L.toUByte()", field.name, combined)
-            ScalarKind.UShort -> body.addStatement("val %L = %L.toUShort()", field.name, combined)
-            ScalarKind.UInt, ScalarKind.ULong -> body.addStatement("val %L = %L", field.name, combined)
-            ScalarKind.Byte -> body.addStatement("val %L = %L.toByte()", field.name, combined)
-            ScalarKind.Short -> body.addStatement("val %L = %L.toShort()", field.name, combined)
-            ScalarKind.Int -> body.addStatement("val %L = %L.toInt()", field.name, combined)
-            ScalarKind.Long -> body.addStatement("val %L = %L.toLong()", field.name, combined)
-            ScalarKind.Float ->
-                body.addStatement("val %L = Float.fromBits(%L.toInt())", field.name, combined)
-            ScalarKind.Double ->
-                body.addStatement("val %L = Double.fromBits(%L.toLong())", field.name, combined)
-            ScalarKind.Boolean ->
-                error("Boolean is pinned to the natural-read path; analyzeField rejects manual-path Boolean")
-        }
-    }
-
-    private fun appendEncodeScalar(
-        body: CodeBlock.Builder,
-        field: FieldSpec.Scalar,
-        ownerSimpleName: String,
-    ) {
-        appendEncodeGuard(body, field, ownerSimpleName)
-        val accessor = "value.${field.name}"
-        val widthMatches = field.wireBytes == field.kind.width
-        val explicitOrder = field.resolvedWireOrder != Endianness.Default
-        if (widthMatches && !explicitOrder) {
-            body.addStatement(naturalScalarWriteStatement(field.kind, accessor))
-            return
-        }
-        // Natural-width explicit Big/Little on a multi-byte scalar. Convert
-        // to the natural integer type, conditionally swapBytes, write.
-        if (widthMatches && explicitOrder && field.kind.width > 1) {
-            appendNaturalWriteWithSwap(body, field, accessor)
-            return
-        }
-        val bigEndian =
-            when (field.resolvedWireOrder) {
-                Endianness.Little -> false
-                Endianness.Big, Endianness.Default -> true
-            }
-        appendManualScalarEncode(body, field, accessor, bigEndian)
-    }
-
-    private fun appendNaturalWriteWithSwap(
-        body: CodeBlock.Builder,
-        field: FieldSpec.Scalar,
-        accessor: String,
-    ) {
-        val canonicalOrder =
-            if (field.resolvedWireOrder == Endianness.Big) "BIG_ENDIAN" else "LITTLE_ENDIAN"
-        val writeMethod =
-            when (field.kind.width) {
-                2 -> "writeShort"
-                4 -> "writeInt"
-                8 -> "writeLong"
-                else -> error("unsupported natural width ${field.kind.width}")
-            }
-        val rawExpr =
-            when (field.kind) {
-                ScalarKind.Short, ScalarKind.Int, ScalarKind.Long -> accessor
-                ScalarKind.UShort -> "$accessor.toShort()"
-                ScalarKind.UInt -> "$accessor.toInt()"
-                ScalarKind.ULong -> "$accessor.toLong()"
-                ScalarKind.Float -> "$accessor.toRawBits()"
-                ScalarKind.Double -> "$accessor.toRawBits()"
-                ScalarKind.UByte, ScalarKind.Byte, ScalarKind.Boolean ->
-                    error("1-byte scalar should not take the natural-write-with-swap path")
-            }
-        val rawVar = "${field.name}Raw"
-        body.addStatement("val %L = %L", rawVar, rawExpr)
-        body.addStatement(
-            "buffer.%L(if (buffer.byteOrder == %T.%L) %L else %M(%L))",
-            writeMethod,
-            BYTE_ORDER_CN,
-            canonicalOrder,
-            rawVar,
-            SWAP_BYTES_MN,
-            rawVar,
-        )
-    }
-
-    private fun appendEncodeGuard(
-        body: CodeBlock.Builder,
-        field: FieldSpec.Scalar,
-        ownerSimpleName: String,
-    ) {
-        if (field.wireBytes >= field.kind.width) return
-        val accessor = "value.${field.name}"
-        val (lhs, maxLit) =
-            when (field.kind) {
-                ScalarKind.ULong -> accessor to "((1uL shl ${8 * field.wireBytes}) - 1uL)"
-                ScalarKind.UInt -> accessor to "((1u shl ${8 * field.wireBytes}) - 1u)"
-                ScalarKind.UShort -> "$accessor.toUInt()" to "((1u shl ${8 * field.wireBytes}) - 1u)"
-                // wireBytes < 1 is rejected by analyzeField
-                ScalarKind.UByte -> return
-                // signed kinds reject @WireBytes narrowing in analyzeField
-                ScalarKind.Byte, ScalarKind.Short, ScalarKind.Int, ScalarKind.Long -> return
-                // Float/Double also reject @WireBytes narrowing
-                ScalarKind.Float, ScalarKind.Double -> return
-                // analyzeField pins Boolean to natural width — never narrows
-                ScalarKind.Boolean -> return
-            }
-        val maxValue = (1L shl (8 * field.wireBytes)) - 1
-        body.beginControlFlow("if (%L > %L)", lhs, maxLit)
-        body.addStatement(
-            "throw %T(fieldPath = %S, reason = %S)",
-            ENCODE_EXCEPTION_CN,
-            "$ownerSimpleName.${field.name}",
-            "value exceeds @WireBytes(${field.wireBytes}) range (max $maxValue)",
-        )
-        body.endControlFlow()
-    }
-
-    private fun appendManualScalarEncode(
-        body: CodeBlock.Builder,
-        field: FieldSpec.Scalar,
-        accessor: String,
-        bigEndian: Boolean,
-    ) {
-        val width = field.wireBytes
-        if (field.kind == ScalarKind.UByte && width == 1) {
-            body.addStatement("buffer.writeUByte(%L)", accessor)
-            return
-        }
-        if (field.kind == ScalarKind.Byte && width == 1) {
-            body.addStatement("buffer.writeByte(%L)", accessor)
-            return
-        }
-        // Convert the field value to a wide unsigned accumulator (UInt for
-        // ≤4 bytes, ULong for >4) and shift bytes off the high or low end
-        // per wire order. Signed kinds reinterpret via .toUInt()/.toULong()
-        // (bit-preserving). Float/Double go through toRawBits().
-        val wide =
-            when (field.kind) {
-                ScalarKind.UByte -> "$accessor.toUInt()"
-                ScalarKind.UShort -> "$accessor.toUInt()"
-                ScalarKind.UInt -> accessor
-                ScalarKind.ULong -> accessor
-                ScalarKind.Byte -> "$accessor.toUByte().toUInt()"
-                ScalarKind.Short -> "$accessor.toUShort().toUInt()"
-                ScalarKind.Int -> "$accessor.toUInt()"
-                ScalarKind.Long -> "$accessor.toULong()"
-                ScalarKind.Float -> "$accessor.toRawBits().toUInt()"
-                ScalarKind.Double -> "$accessor.toRawBits().toULong()"
-                ScalarKind.Boolean ->
-                    error("Boolean is pinned to the natural-write path; analyzeField rejects manual-path Boolean")
-            }
-        val maskLit = if (width >= 5) "0xFFuL" else "0xFFu"
-        for (i in 0 until width) {
-            val shiftBits = if (bigEndian) (width - 1 - i) * 8 else i * 8
-            val expr =
-                if (shiftBits == 0) {
-                    "$wide and $maskLit"
-                } else {
-                    "($wide shr $shiftBits) and $maskLit"
-                }
-            body.addStatement("buffer.writeUByte((%L).toUByte())", expr)
-        }
-    }
-
-    /**
-     * Emit decode for a `@JvmInline value class` field
-     * with a single supported-scalar inner. Reads the inner scalar at
-     * natural width and constructs the value class via its primary
-     * constructor. The local is named after the outer parameter so
-     * dotted-form `@When` resolvers can address it as `<name>.<property>`.
-     */
-    private fun appendDecodeValueClassScalar(
-        body: CodeBlock.Builder,
-        field: FieldSpec.ValueClassScalar,
-    ) {
-        // Honor the value class's declared @ProtocolMessage(wireOrder) the
-        // same way plain Scalars honor @WireOrder / parent wireOrder: explicit
-        // Big / Little wins over buffer.byteOrder. Multi-byte inner kinds get
-        // the swapBytes fast path (matches the single-scalar Scalar emit).
-        // 1-byte inner kinds have no byte order — the natural read suffices.
-        if (field.valueClassWireOrder != Endianness.Default && field.innerKind.width > 1) {
-            appendValueClassNaturalReadWithSwap(body, field)
-            return
-        }
-        body.addStatement(
-            "val %L = %T(%L)",
-            field.name,
-            field.valueClassType,
-            naturalScalarReadExpr(field.innerKind),
-        )
-    }
-
-    /**
-     * Emit encode for a value-class field. Unwraps
-     * via the inner property name and writes the inner scalar at
-     * natural width — or, when the value class declares an explicit
-     * wireOrder, takes the swap fast path so the wire bytes match
-     * the value class's contract regardless of buffer.byteOrder.
-     */
-    private fun appendEncodeValueClassScalar(
-        body: CodeBlock.Builder,
-        field: FieldSpec.ValueClassScalar,
-    ) {
-        if (field.valueClassWireOrder != Endianness.Default && field.innerKind.width > 1) {
-            appendValueClassNaturalWriteWithSwap(body, field)
-            return
-        }
-        body.addStatement(
-            naturalScalarWriteStatement(
-                field.innerKind,
-                "value.${field.name}.${field.innerPropertyName}",
-            ),
-        )
-    }
-
-    private fun appendValueClassNaturalReadWithSwap(
-        body: CodeBlock.Builder,
-        field: FieldSpec.ValueClassScalar,
-    ) {
-        val canonicalOrder =
-            if (field.valueClassWireOrder == Endianness.Big) "BIG_ENDIAN" else "LITTLE_ENDIAN"
-        val readMethod =
-            when (field.innerKind.width) {
-                2 -> "readShort"
-                4 -> "readInt"
-                8 -> "readLong"
-                else -> error("unsupported value-class inner width ${field.innerKind.width}")
-            }
-        val rawVar = "${field.name}Raw"
-        body.addStatement("val %L = buffer.%L()", rawVar, readMethod)
-        val toUnsigned =
-            when (field.innerKind) {
-                ScalarKind.UShort -> "toUShort"
-                ScalarKind.UInt -> "toUInt"
-                ScalarKind.ULong -> "toULong"
-                else -> null
-            }
-        if (toUnsigned != null) {
-            body.addStatement(
-                "val %L = %T((if (buffer.byteOrder == %T.%L) %L else %M(%L)).%L())",
-                field.name,
-                field.valueClassType,
-                BYTE_ORDER_CN,
-                canonicalOrder,
-                rawVar,
-                SWAP_BYTES_MN,
-                rawVar,
-                toUnsigned,
-            )
-        } else {
-            body.addStatement(
-                "val %L = %T(if (buffer.byteOrder == %T.%L) %L else %M(%L))",
-                field.name,
-                field.valueClassType,
-                BYTE_ORDER_CN,
-                canonicalOrder,
-                rawVar,
-                SWAP_BYTES_MN,
-                rawVar,
-            )
-        }
-    }
-
-    private fun appendValueClassNaturalWriteWithSwap(
-        body: CodeBlock.Builder,
-        field: FieldSpec.ValueClassScalar,
-    ) {
-        val canonicalOrder =
-            if (field.valueClassWireOrder == Endianness.Big) "BIG_ENDIAN" else "LITTLE_ENDIAN"
-        val writeMethod =
-            when (field.innerKind.width) {
-                2 -> "writeShort"
-                4 -> "writeInt"
-                8 -> "writeLong"
-                else -> error("unsupported value-class inner width ${field.innerKind.width}")
-            }
-        val accessor = "value.${field.name}.${field.innerPropertyName}"
-        val rawExpr =
-            when (field.innerKind) {
-                ScalarKind.Short, ScalarKind.Int, ScalarKind.Long -> accessor
-                ScalarKind.UShort -> "$accessor.toShort()"
-                ScalarKind.UInt -> "$accessor.toInt()"
-                ScalarKind.ULong -> "$accessor.toLong()"
-                ScalarKind.Float -> "$accessor.toRawBits()"
-                ScalarKind.Double -> "$accessor.toRawBits()"
-                else -> error("inner kind ${field.innerKind} cannot reach the swap path")
-            }
-        val rawVar = "${field.name}Raw"
-        body.addStatement("val %L = %L", rawVar, rawExpr)
-        body.addStatement(
-            "buffer.%L(if (buffer.byteOrder == %T.%L) %L else %M(%L))",
-            writeMethod,
-            BYTE_ORDER_CN,
-            canonicalOrder,
-            rawVar,
-            SWAP_BYTES_MN,
-            rawVar,
-        )
-    }
-
-    /**
-     * Emit a `@When` decode block.
-     *
-     * Generated shape:
-     * ```
-     * val <name>: <NullableType> = if (<source>) <readExpr> else null
-     * ```
-     *
-     * The source is a sibling `Boolean` local already in scope (decode visits
-     * fields in constructor order, and analyzeConditionalField has verified
-     * the source is declared before this field). `readExpr` is the natural-
-     * width scalar read for the inner kind ( restricts inner to a
-     * natural-width Scalar; widens to LengthPrefixedString).
-     */
-    private fun appendDecodeConditional(
-        body: CodeBlock.Builder,
-        field: FieldSpec.Conditional,
-    ) {
-        when (val inner = field.inner) {
-            is ConditionalInner.Scalar -> {
-                if (inner.wireOrder != Endianness.Default && inner.kind.width > 1) {
-                    appendConditionalScalarSwapDecode(
-                        body = body,
-                        fieldName = field.name,
-                        nullableTypeName = field.nullableTypeName,
-                        condition = decodeConditionExpr(field.condition),
-                        kind = inner.kind,
-                        wireOrder = inner.wireOrder,
-                        wrapValueClass = null,
-                    )
-                } else {
-                    body.addStatement(
-                        "val %L: %T = if (%L) %L else null",
-                        field.name,
-                        field.nullableTypeName,
-                        decodeConditionExpr(field.condition),
-                        naturalScalarReadExpr(inner.kind),
-                    )
-                }
-            }
-            is ConditionalInner.LengthPrefixedString -> {
-                body.beginControlFlow(
-                    "val %L: %T = if (%L)",
-                    field.name,
-                    field.nullableTypeName,
-                    decodeConditionExpr(field.condition),
-                )
-                val lengthVar =
-                    appendLengthPrefixedStringPrefixDecode(
-                        body = body,
-                        name = field.name,
-                        ownerSimpleName = field.ownerSimpleName,
-                        prefixWidth = inner.prefixWidth,
-                        prefixWireOrder = inner.prefixWireOrder,
-                    )
-                body.addStatement("buffer.readString(%L, %T.UTF8)", lengthVar, CHARSET_CN)
-                body.nextControlFlow("else")
-                body.addStatement("null")
-                body.endControlFlow()
-            }
-            is ConditionalInner.ValueClassScalar -> {
-                // Wrap the natural-width inner read
-                // in the value-class constructor (mirror of 's
-                // non-conditional `appendDecodeValueClassScalar`).
-                if (inner.valueClassWireOrder != Endianness.Default && inner.innerKind.width > 1) {
-                    appendConditionalScalarSwapDecode(
-                        body = body,
-                        fieldName = field.name,
-                        nullableTypeName = field.nullableTypeName,
-                        condition = decodeConditionExpr(field.condition),
-                        kind = inner.innerKind,
-                        wireOrder = inner.valueClassWireOrder,
-                        wrapValueClass = inner.valueClassType,
-                    )
-                } else {
-                    body.addStatement(
-                        "val %L: %T = if (%L) %T(%L) else null",
-                        field.name,
-                        field.nullableTypeName,
-                        decodeConditionExpr(field.condition),
-                        inner.valueClassType,
-                        naturalScalarReadExpr(inner.innerKind),
-                    )
-                }
-            }
-            is ConditionalInner.LengthPrefixedUseCodecList -> {
-                // `@When @LengthPrefixed @UseCodec(C) val
-                // xs: List<E>?` — predicate-true branch runs the
-                // inner-bag decode ( shared body). Else null.
-                body.beginControlFlow(
-                    "val %L: %T = if (%L)",
-                    field.name,
-                    field.nullableTypeName,
-                    decodeConditionExpr(field.condition),
-                )
-                appendDecodeLengthPrefixedListBody(
-                    body = body,
-                    spec = inner.spec,
-                    listLocalName = "${field.name}Value",
-                    namespacePrefix = field.name,
-                )
-                body.addStatement("%LValue", field.name)
-                body.nextControlFlow("else")
-                body.addStatement("null")
-                body.endControlFlow()
-            }
-            is ConditionalInner.LengthPrefixedUseCodecPayload -> {
-                // Predicate-true branch reads the
-                // fixed-width prefix, narrows `buffer.limit()` to position
-                // + length, runs `<C>.decode`, restores the outer limit.
-                // Mirrors [appendDecodeLengthPrefixedUseCodecPayload] but
-                // wrapped in the conditional's `if (predicate)` gate.
-                body.beginControlFlow(
-                    "val %L: %T = if (%L)",
-                    field.name,
-                    field.nullableTypeName,
-                    decodeConditionExpr(field.condition),
-                )
-                val lengthVar =
-                    appendLengthPrefixedStringPrefixDecode(
-                        body = body,
-                        name = field.name,
-                        ownerSimpleName = field.ownerSimpleName,
-                        prefixWidth = inner.prefixWidth,
-                        prefixWireOrder = inner.prefixWireOrder,
-                    )
-                val outerLimitVar = "__${field.name}OuterLimit"
-                body.addStatement("val %L = buffer.limit()", outerLimitVar)
-                body.addStatement("buffer.setLimit(buffer.position() + %L)", lengthVar)
-                body.beginControlFlow("try")
-                body.addStatement("%T.decode(buffer, context)", inner.payloadCodecType)
-                body.nextControlFlow("finally")
-                body.addStatement("buffer.setLimit(%L)", outerLimitVar)
-                body.endControlFlow()
-                body.nextControlFlow("else")
-                body.addStatement("null")
-                body.endControlFlow()
-            }
-            is ConditionalInner.UseCodecScalar -> {
-                // `@When @UseCodec(C) val: T?`.
-                // Predicate-true delegates to the codec object's
-                // `decode(buffer, context)`, just like the non-conditional
-                // `appendDecodeUseCodecScalar` path; predicate-false yields
-                // null. The cascading-trailer cases use grammar-2
-                // `remaining >= N` predicates so the read only runs when
-                // the bounded buffer still has bytes to spend.
-                body.addStatement(
-                    "val %L: %T = if (%L) %T.decode(buffer, context) else null",
-                    field.name,
-                    field.nullableTypeName,
-                    decodeConditionExpr(field.condition),
-                    inner.codecType,
-                )
-            }
-            is ConditionalInner.ProtocolMessageScalar -> {
-                // Bare `@When val: T?` for a
-                // `@ProtocolMessage` data class or sealed parent. The
-                // codec class resolves to `${T.simpleName}Codec`
-                // by-name; the call shape is identical to UseCodecScalar.
-                body.addStatement(
-                    "val %L: %T = if (%L) %T.decode(buffer, context) else null",
-                    field.name,
-                    field.nullableTypeName,
-                    decodeConditionExpr(field.condition),
-                    inner.codecType,
-                )
-            }
-        }
-    }
-
-    /**
-     * /3 — emit a `@When` encode block.
-     *
-     * Generated shape:
-     * ```
-     * if (value.<source>) {
-     *     val <name>Value = value.<name> ?: throw EncodeException(...)
-     *     <writeStatement(s) using `<name>Value`>
-     * }
-     * ```
-     *
-     * `<source>` is `value.<sibling>` for the simple form and
-     * `value.<sibling>.<property>` for the dotted form. The body is
-     * a single-line scalar write for `ConditionalInner.Scalar` and
-     * the BackPatch length-prefix sequence for
-     * `ConditionalInner.LengthPrefixedString` (.5).
-     *
-     * Predicate-false branch writes nothing (zero bytes for the slot, per
-     * ). Predicate-true with `value.<name> == null`
-     * throws `EncodeException` with field-path attribution (row 20).
-     */
-    private fun appendEncodeConditional(
-        body: CodeBlock.Builder,
-        field: FieldSpec.Conditional,
-    ) {
-        body.beginControlFlow("if (%L)", encodeConditionAccessor(field.condition, field.name))
-        val localName = "${field.name}Value"
-        body.addStatement(
-            "val %L = value.%L ?: throw %T(fieldPath = %S, reason = %S)",
-            localName,
-            field.name,
-            ENCODE_EXCEPTION_CN,
-            "${field.ownerSimpleName}.${field.name}",
-            "@When(\"${conditionExpressionLiteral(field.condition)}\") predicate is true but field is null",
-        )
-        when (val inner = field.inner) {
-            is ConditionalInner.Scalar ->
-                if (inner.wireOrder != Endianness.Default && inner.kind.width > 1) {
-                    appendConditionalScalarSwapEncode(
-                        body = body,
-                        accessor = localName,
-                        kind = inner.kind,
-                        wireOrder = inner.wireOrder,
-                        valueClassInnerProperty = null,
-                    )
-                } else {
-                    body.addStatement(naturalScalarWriteStatement(inner.kind, localName))
-                }
-            is ConditionalInner.LengthPrefixedString ->
-                appendLengthPrefixedStringEncode(
-                    body = body,
-                    name = field.name,
-                    ownerSimpleName = field.ownerSimpleName,
-                    prefixWidth = inner.prefixWidth,
-                    prefixWireOrder = inner.prefixWireOrder,
-                    accessor = localName,
-                )
-            is ConditionalInner.ValueClassScalar ->
-                // Unwrap the value class via the
-                // inner property name (mirror of 's
-                // non-conditional `appendEncodeValueClassScalar`).
-                if (inner.valueClassWireOrder != Endianness.Default && inner.innerKind.width > 1) {
-                    appendConditionalScalarSwapEncode(
-                        body = body,
-                        accessor = localName,
-                        kind = inner.innerKind,
-                        wireOrder = inner.valueClassWireOrder,
-                        valueClassInnerProperty = inner.innerPropertyName,
-                    )
-                } else {
-                    body.addStatement(
-                        naturalScalarWriteStatement(
-                            inner.innerKind,
-                            "$localName.${inner.innerPropertyName}",
-                        ),
-                    )
-                }
-            is ConditionalInner.LengthPrefixedUseCodecList ->
-                appendEncodeConditionalLengthPrefixedUseCodecList(
-                    body = body,
-                    field = field,
-                    inner = inner,
-                    accessor = localName,
-                )
-            is ConditionalInner.LengthPrefixedUseCodecPayload ->
-                // BackPatch shape mirroring
-                // [appendEncodeLengthPrefixedUseCodecPayload]: reserve
-                // prefix slot, run `<C>.encode`, measure body byte count,
-                // patch the prefix in place, restore position. Reads the
-                // smart-cast non-null `<name>Value` local established by
-                // the outer `appendEncodeConditional`.
-                appendEncodeConditionalLengthPrefixedUseCodecPayload(
-                    body = body,
-                    field = field,
-                    inner = inner,
-                    accessor = localName,
-                )
-            is ConditionalInner.UseCodecScalar ->
-                // Mirror of the non-conditional
-                // `appendEncodeUseCodecScalar`. Predicate-true with
-                // smart-cast non-null `<name>Value` (established above)
-                // delegates to the user codec's `encode`.
-                body.addStatement(
-                    "%T.encode(buffer, %L, context)",
-                    inner.codecType,
-                    localName,
-                )
-            is ConditionalInner.ProtocolMessageScalar ->
-                // Same encode shape as
-                // UseCodecScalar; the only thing that differs is how the
-                // codec class name was resolved at analyze time.
-                body.addStatement(
-                    "%T.encode(buffer, %L, context)",
-                    inner.codecType,
-                    localName,
-                )
-        }
-        body.endControlFlow()
-    }
-
-    // Shared helper for the @When + explicit-wireOrder case. Generates a
-    // block-expression if/else where the `if` arm reads the natural-width
-    // wire value and canonicalizes via swapBytes; matches the contract
-    // explicit wire order should beat buffer.byteOrder, mirroring the
-    // non-conditional Scalar / ValueClassScalar swap path. `wrapValueClass`
-    // routes the swapped value through the value class's constructor when
-    // present (ValueClassScalar conditional path).
-    private fun appendConditionalScalarSwapDecode(
-        body: CodeBlock.Builder,
-        fieldName: String,
-        nullableTypeName: TypeName,
-        condition: String,
-        kind: ScalarKind,
-        wireOrder: Endianness,
-        wrapValueClass: ClassName?,
-    ) {
-        val canonicalOrder =
-            if (wireOrder == Endianness.Big) "BIG_ENDIAN" else "LITTLE_ENDIAN"
-        val readMethod =
-            when (kind.width) {
-                2 -> "readShort"
-                4 -> "readInt"
-                8 -> "readLong"
-                else -> error("unsupported conditional width ${kind.width}")
-            }
-        val rawVar = "${fieldName}Raw"
-        body.beginControlFlow(
-            "val %L: %T = if (%L)",
-            fieldName,
-            nullableTypeName,
-            condition,
-        )
-        body.addStatement("val %L = buffer.%L()", rawVar, readMethod)
-        // Emit the swap + cast (+ optional value-class wrap) as the if-block's
-        // value expression via KotlinPoet placeholders so ByteOrder and
-        // swapBytes resolve through the file's imports (not as FQNs).
-        val unsignedCast =
-            when (kind) {
-                ScalarKind.UShort -> "toUShort"
-                ScalarKind.UInt -> "toUInt"
-                ScalarKind.ULong -> "toULong"
-                else -> null
-            }
-        when {
-            wrapValueClass != null && unsignedCast != null ->
-                body.addStatement(
-                    "%T((if (buffer.byteOrder == %T.%L) %L else %M(%L)).%L())",
-                    wrapValueClass,
-                    BYTE_ORDER_CN,
-                    canonicalOrder,
-                    rawVar,
-                    SWAP_BYTES_MN,
-                    rawVar,
-                    unsignedCast,
-                )
-            wrapValueClass != null ->
-                body.addStatement(
-                    "%T(if (buffer.byteOrder == %T.%L) %L else %M(%L))",
-                    wrapValueClass,
-                    BYTE_ORDER_CN,
-                    canonicalOrder,
-                    rawVar,
-                    SWAP_BYTES_MN,
-                    rawVar,
-                )
-            unsignedCast != null ->
-                body.addStatement(
-                    "(if (buffer.byteOrder == %T.%L) %L else %M(%L)).%L()",
-                    BYTE_ORDER_CN,
-                    canonicalOrder,
-                    rawVar,
-                    SWAP_BYTES_MN,
-                    rawVar,
-                    unsignedCast,
-                )
-            kind == ScalarKind.Float ->
-                body.addStatement(
-                    "Float.fromBits(if (buffer.byteOrder == %T.%L) %L else %M(%L))",
-                    BYTE_ORDER_CN,
-                    canonicalOrder,
-                    rawVar,
-                    SWAP_BYTES_MN,
-                    rawVar,
-                )
-            kind == ScalarKind.Double ->
-                body.addStatement(
-                    "Double.fromBits(if (buffer.byteOrder == %T.%L) %L else %M(%L))",
-                    BYTE_ORDER_CN,
-                    canonicalOrder,
-                    rawVar,
-                    SWAP_BYTES_MN,
-                    rawVar,
-                )
-            else ->
-                body.addStatement(
-                    "if (buffer.byteOrder == %T.%L) %L else %M(%L)",
-                    BYTE_ORDER_CN,
-                    canonicalOrder,
-                    rawVar,
-                    SWAP_BYTES_MN,
-                    rawVar,
-                )
-        }
-        body.nextControlFlow("else")
-        body.addStatement("null")
-        body.endControlFlow()
-    }
-
-    private fun appendConditionalScalarSwapEncode(
-        body: CodeBlock.Builder,
-        accessor: String,
-        kind: ScalarKind,
-        wireOrder: Endianness,
-        valueClassInnerProperty: String?,
-    ) {
-        val canonicalOrder =
-            if (wireOrder == Endianness.Big) "BIG_ENDIAN" else "LITTLE_ENDIAN"
-        val writeMethod =
-            when (kind.width) {
-                2 -> "writeShort"
-                4 -> "writeInt"
-                8 -> "writeLong"
-                else -> error("unsupported conditional width ${kind.width}")
-            }
-        val resolvedAccessor =
-            if (valueClassInnerProperty != null) "$accessor.$valueClassInnerProperty" else accessor
-        val rawExpr =
-            when (kind) {
-                ScalarKind.Short, ScalarKind.Int, ScalarKind.Long -> resolvedAccessor
-                ScalarKind.UShort -> "$resolvedAccessor.toShort()"
-                ScalarKind.UInt -> "$resolvedAccessor.toInt()"
-                ScalarKind.ULong -> "$resolvedAccessor.toLong()"
-                ScalarKind.Float -> "$resolvedAccessor.toRawBits()"
-                ScalarKind.Double -> "$resolvedAccessor.toRawBits()"
-                ScalarKind.UByte, ScalarKind.Byte, ScalarKind.Boolean ->
-                    error("1-byte kind should not take the conditional swap path")
-            }
-        val rawVar = "${accessor}Raw"
-        body.addStatement("val %L = %L", rawVar, rawExpr)
-        body.addStatement(
-            "buffer.%L(if (buffer.byteOrder == %T.%L) %L else %M(%L))",
-            writeMethod,
-            BYTE_ORDER_CN,
-            canonicalOrder,
-            rawVar,
-            SWAP_BYTES_MN,
-            rawVar,
-        )
-    }
-
-    /**
-     * Encode a conditional `@LengthPrefixed @UseCodec(C)
-     * val xs: List<E>?`. Audit-2a deduplication: delegates to the shared
-     * `appendEncodeLengthPrefixedListBody` helper.
-     *
-     * `accessor` is the smart-cast non-null local established by the
-     * outer `appendEncodeConditional` (`<name>Value`). The
-     * non-conditional emit reads `value.<name>` instead — same shape,
-     * different read expression (the helper takes `accessor` as a
-     * parameter to absorb the difference).
-     */
-    private fun appendEncodeConditionalLengthPrefixedUseCodecList(
-        body: CodeBlock.Builder,
-        field: FieldSpec.Conditional,
-        inner: ConditionalInner.LengthPrefixedUseCodecList,
-        accessor: String,
-    ) {
-        appendEncodeLengthPrefixedListBody(
-            body = body,
-            spec = inner.spec,
-            accessor = accessor,
-            namespacePrefix = field.name,
-        )
-    }
-
-    /**
-     * Encode a conditional `@LengthPrefixed
-     * @UseCodec(C) val: T?` where T : Payload. BackPatch shape mirroring
-     * [appendEncodeLengthPrefixedUseCodecPayload]: reserve prefix slot,
-     * run `<C>.encode(buffer, accessor, context)` against the
-     * accumulating buffer, measure body byte count from the position
-     * delta, patch the prefix, restore position past the body.
-     *
-     * `accessor` is the smart-cast non-null local established by the
-     * outer `appendEncodeConditional` (`<name>Value`). The non-
-     * conditional emit reads `value.<name>` directly.
-     */
-    private fun appendEncodeConditionalLengthPrefixedUseCodecPayload(
-        body: CodeBlock.Builder,
-        field: FieldSpec.Conditional,
-        inner: ConditionalInner.LengthPrefixedUseCodecPayload,
-        accessor: String,
-    ) {
-        val sizePosVar = "${field.name}SizePosition"
-        val bodyStartVar = "${field.name}BodyStart"
-        val endPosVar = "${field.name}EndPosition"
-        val byteCountVar = "${field.name}ByteCount"
-        body.addStatement("val %L = buffer.position()", sizePosVar)
-        body.addStatement("buffer.position(%L + %L)", sizePosVar, inner.prefixWidth)
-        body.addStatement("val %L = buffer.position()", bodyStartVar)
-        body.addStatement("%T.encode(buffer, %L, context)", inner.payloadCodecType, accessor)
-        body.addStatement("val %L = buffer.position()", endPosVar)
-        body.addStatement("val %L = %L - %L", byteCountVar, endPosVar, bodyStartVar)
-        if (inner.prefixWidth < 4) {
-            val maxValue = (1L shl (inner.prefixWidth * 8)) - 1
-            val widthName =
-                when (inner.prefixWidth) {
-                    1 -> "Byte"
-                    2 -> "Short"
-                    else -> error("unreachable: prefixWidth must be 1, 2, or 4")
-                }
-            body.beginControlFlow("if (%L > %L)", byteCountVar, maxValue)
-            body.addStatement(
-                "throw %T(fieldPath = %S, reason = %P)",
-                ENCODE_EXCEPTION_CN,
-                "${field.ownerSimpleName}.${field.name}",
-                "encoded payload byte length \${$byteCountVar} exceeds " +
-                    "@LengthPrefixed(LengthPrefix.$widthName) max $maxValue",
-            )
-            body.endControlFlow()
-        }
-        body.addStatement("buffer.position(%L)", sizePosVar)
-        val prefixVar = "${field.name}Prefix"
-        body.addStatement("val %L = %L.toUInt()", prefixVar, byteCountVar)
-        appendBufferPrefixEncode(body, prefixVar, inner.prefixWidth, inner.prefixWireOrder)
-        body.addStatement("buffer.position(%L)", endPosVar)
-    }
-
-    /**
-     * Encode-side predicate accessor. Encode reads from the message
-     * value, so all paths start at `value.`. Simple form is
-     * `value.<sibling>`; dotted form is `value.<sibling>.<property>`.
-     *
-     * Grammar 2 (`remaining <op> <int>`) encode semantics differ:
-     * cascading-trailer fields are gated on whether the caller provided
-     * a non-null value (the encode-side has no buffer-`remaining()` to
-     * test against — the slot is included iff the field is set). Caller
-     * is responsible for keeping the cascade consistent (don't set a
-     * later trailer if an earlier one is null).
-     */
-    private fun encodeConditionAccessor(
-        condition: ConditionRef,
-        fieldName: String,
-    ): String =
-        when (condition) {
-            is ConditionRef.Sibling -> "value.${condition.name}"
-            is ConditionRef.ValueClassProperty -> "value.${condition.siblingName}.${condition.propertyName}"
-            is ConditionRef.RemainingCmp -> "value.$fieldName != null"
-        }
-
-    /**
-     * Reconstruct the original `@When("...")` expression literal
-     * for use in `EncodeException` field-path messages (row 20).
-     */
-    private fun conditionExpressionLiteral(condition: ConditionRef): String =
-        when (condition) {
-            is ConditionRef.Sibling -> condition.name
-            is ConditionRef.ValueClassProperty -> "${condition.siblingName}.${condition.propertyName}"
-            is ConditionRef.RemainingCmp ->
-                "remaining ${condition.op.symbol} ${condition.threshold}"
-        }
-
-    private fun appendDecodeLengthPrefixed(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthPrefixedMessage,
-    ) {
-        val prefixVar = "${field.name}Prefix"
-        appendBufferPrefixDecode(body, prefixVar, field.prefixWidth, field.prefixWireOrder)
-        body.beginControlFlow("if (%L > Int.MAX_VALUE.toUInt())", prefixVar)
-        body.addStatement(
-            "throw %T(fieldPath = %S, bufferPosition = -1, expected = %S, actual = %L.toString())",
-            DECODE_EXCEPTION_CN,
-            "${field.ownerSimpleName}.${field.name}",
-            "length prefix <= \${Int.MAX_VALUE}",
-            prefixVar,
-        )
-        body.endControlFlow()
-        val resolvedVar = "${field.name}Length"
-        body.addStatement("val %L = %L.toInt()", resolvedVar, prefixVar)
-        val outerVar = "${field.name}OuterLimit"
-        body.addStatement("val %L = buffer.limit()", outerVar)
-        body.addStatement("buffer.setLimit(buffer.position() + %L)", resolvedVar)
-        body.beginControlFlow("val %L = try", field.name)
-        body.addStatement("%T.decode(buffer, context)", field.codecType)
-        body.nextControlFlow("finally")
-        body.addStatement("buffer.setLimit(%L)", outerVar)
-        body.endControlFlow()
-    }
-
-    private fun appendBufferPrefixDecode(
-        body: CodeBlock.Builder,
-        targetVar: String,
-        prefixWidth: Int,
-        wireOrder: Endianness,
-    ) {
-        if (prefixWidth == 1) {
-            body.addStatement("val %L = buffer.readUByte().toUInt()", targetVar)
-            return
-        }
-        val bigEndian =
-            when (wireOrder) {
-                Endianness.Big -> true
-                Endianness.Little -> false
-                Endianness.Default -> true
-            }
-        for (i in 0 until prefixWidth) {
-            body.addStatement(
-                "val %L = buffer.readUByte().toUInt()",
-                "${targetVar}B$i",
-            )
-        }
-        val parts =
-            (0 until prefixWidth).map { i ->
-                val byteName = "${targetVar}B$i"
-                val shiftBits = if (bigEndian) (prefixWidth - 1 - i) * 8 else i * 8
-                if (shiftBits == 0) byteName else "($byteName shl $shiftBits)"
-            }
-        body.addStatement("val %L = (%L)", targetVar, parts.joinToString(" or "))
-    }
-
-    private fun appendEncodeLengthPrefixed(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthPrefixedMessage,
-    ) {
-        val prefixVar = "${field.name}Prefix"
-        body.addStatement(
-            "val %L = (%T.wireSize(value.%L, context) as %T.Exact).bytes.toUInt()",
-            prefixVar,
-            field.codecType,
-            field.name,
-            WIRE_SIZE_CN,
-        )
-        appendBufferPrefixEncode(body, prefixVar, field.prefixWidth, field.prefixWireOrder)
-        body.addStatement("%T.encode(buffer, value.%L, context)", field.codecType, field.name)
-    }
-
-    private fun appendDecodeLengthPrefixedString(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthPrefixedString,
-    ) {
-        val lengthVar =
-            appendLengthPrefixedStringPrefixDecode(
-                body = body,
-                name = field.name,
-                ownerSimpleName = field.ownerSimpleName,
-                prefixWidth = field.prefixWidth,
-                prefixWireOrder = field.prefixWireOrder,
-            )
-        body.addStatement(
-            "val %L = buffer.readString(%L, %T.UTF8)",
-            field.name,
-            lengthVar,
-            CHARSET_CN,
-        )
-    }
-
-    /**
-     * Emit the prefix read + Int.MAX_VALUE guard + length
-     * Int conversion shared by length-prefixed-string field decode
-     * and the conditional `@LengthPrefixed @When` decode path.
-     * Returns the local variable name holding the resolved
-     * (Int-typed) length.
-     */
-    private fun appendLengthPrefixedStringPrefixDecode(
-        body: CodeBlock.Builder,
-        name: String,
-        ownerSimpleName: String,
-        prefixWidth: Int,
-        prefixWireOrder: Endianness,
-    ): String {
-        val prefixVar = "${name}Prefix"
-        appendBufferPrefixDecode(body, prefixVar, prefixWidth, prefixWireOrder)
-        body.beginControlFlow("if (%L > Int.MAX_VALUE.toUInt())", prefixVar)
-        body.addStatement(
-            "throw %T(fieldPath = %S, bufferPosition = -1, expected = %S, actual = %L.toString())",
-            DECODE_EXCEPTION_CN,
-            "$ownerSimpleName.$name",
-            "length prefix <= \${Int.MAX_VALUE}",
-            prefixVar,
-        )
-        body.endControlFlow()
-        val lengthVar = "${name}Length"
-        body.addStatement("val %L = %L.toInt()", lengthVar, prefixVar)
-        return lengthVar
-    }
-
-    private fun appendEncodeLengthPrefixedString(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthPrefixedString,
-    ) {
-        appendLengthPrefixedStringEncode(
-            body = body,
-            name = field.name,
-            ownerSimpleName = field.ownerSimpleName,
-            prefixWidth = field.prefixWidth,
-            prefixWireOrder = field.prefixWireOrder,
-            accessor = "value.${field.name}",
-        )
-    }
-
-    /**
-     * Shared BackPatch encoder for length-prefixed-string
-     * fields and the conditional `@LengthPrefixed @When` encode
-     * path.
-     *
-     * `accessor` is the expression that yields the string value;
-     * field-form callers pass `value.<name>`, conditional-form
-     * callers pass the locally-bound non-null value (already
-     * null-checked at the conditional gate). `name` is used for
-     * generated-variable naming and the field-path attribution
-     * literal.
-     */
-    private fun appendLengthPrefixedStringEncode(
-        body: CodeBlock.Builder,
-        name: String,
-        ownerSimpleName: String,
-        prefixWidth: Int,
-        prefixWireOrder: Endianness,
-        accessor: String,
-    ) {
-        // BackPatch pattern: reserve prefix slot, write
-        // the body via the runtime's UTF-8 path, measure byte count from the
-        // position delta, patch the prefix in place, restore position past the
-        // body. The runtime's `writeString(text, Charset.UTF8)` is zero-`ByteArray`
-        // on JVM / Apple / JS; the WASM and nonJvm `writeString` paths still
-        // allocate one ByteArray per call (, deferred to a
-        // separate runtime task).
-        val sizePosVar = "${name}SizePosition"
-        val bodyStartVar = "${name}BodyStart"
-        val endPosVar = "${name}EndPosition"
-        val byteCountVar = "${name}ByteCount"
-        body.addStatement("val %L = buffer.position()", sizePosVar)
-        body.addStatement("buffer.position(%L + %L)", sizePosVar, prefixWidth)
-        body.addStatement("val %L = buffer.position()", bodyStartVar)
-        body.addStatement(
-            "buffer.writeString(%L, %T.UTF8)",
-            accessor,
-            CHARSET_CN,
-        )
-        body.addStatement("val %L = buffer.position()", endPosVar)
-        body.addStatement("val %L = %L - %L", byteCountVar, endPosVar, bodyStartVar)
-        // Runtime overflow guard. For 4-byte prefixes the max (UInt.MAX_VALUE =
-        // 2^32-1) exceeds Int.MAX_VALUE, so a position-delta byte count can never
-        // overflow it — the check would be dead code.
-        if (prefixWidth < 4) {
-            val maxValue = (1L shl (prefixWidth * 8)) - 1
-            val widthName =
-                when (prefixWidth) {
-                    1 -> "Byte"
-                    2 -> "Short"
-                    else -> error("unreachable: prefixWidth must be 1, 2, or 4")
-                }
-            body.beginControlFlow("if (%L > %L)", byteCountVar, maxValue)
-            body.addStatement(
-                "throw %T(fieldPath = %S, reason = %P)",
-                ENCODE_EXCEPTION_CN,
-                "$ownerSimpleName.$name",
-                "UTF-8 byte length \${$byteCountVar} exceeds @LengthPrefixed(LengthPrefix.$widthName) max $maxValue",
-            )
-            body.endControlFlow()
-        }
-        body.addStatement("buffer.position(%L)", sizePosVar)
-        val prefixVar = "${name}Prefix"
-        body.addStatement("val %L = %L.toUInt()", prefixVar, byteCountVar)
-        appendBufferPrefixEncode(body, prefixVar, prefixWidth, prefixWireOrder)
-        body.addStatement("buffer.position(%L)", endPosVar)
-    }
-
-    /**
-     * Emit decode for `@LengthFrom("siblingField")
-     * val: String`. The sibling local is in scope (decode visits
-     * fields in constructor order, and analyzeLengthFromStringField
-     * has verified the sibling is declared before this field).
-     *
-     * Generated shape:
-     * ```
-     * <Int.MAX_VALUE guard for sibling kinds whose range exceeds Int>
-     * val <name>Length = <sibling>.toInt()
-     * val <name> = buffer.readString(<name>Length, Charset.UTF8)
-     * ```
-     *
-     * The guard is skipped for `Byte` / `Short` / `Int` / `UByte` /
-     * `UShort`, whose values fit in a non-negative `Int`. `UInt`,
-     * `ULong`, and `Long` need the runtime guard because their range
-     * exceeds `Int.MAX_VALUE`.
-     */
-    private fun appendDecodeLengthFromString(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthFromString,
-    ) {
-        // Simple form needs an Int.MAX_VALUE guard for kinds whose
-        // range exceeds Int (UInt / ULong / Long); the dotted form's
-        // property returns Int directly so no guard is needed.
-        if (field.source is LengthSource.Sibling) {
-            appendLengthFromIntMaxGuard(
-                body = body,
-                siblingAccessor = field.source.siblingName,
-                siblingKind = field.source.siblingKind,
-                ownerSimpleName = field.ownerSimpleName,
-                fieldName = field.name,
-            )
-        }
-        // Inline the sibling/property accessor rather than binding
-        // an intermediate. A `${field.name}Length` intermediate would
-        // shadow the sibling local when the user names the carrier
-        // `<bound>Length` — a natural Kotlin convention that the
-        // generated code must not break.
-        body.addStatement(
-            "val %L = buffer.readString(%L, %T.UTF8)",
-            field.name,
-            field.source.decodeAccessor(),
-            CHARSET_CN,
-        )
-    }
-
-    /**
-     * Emit encode for `@LengthFrom("siblingField")
-     * val: String`. The sibling field has already been encoded by
-     * the prior field's emit step; this step writes only the body.
-     * The user is responsible for keeping `value.<sibling>`
-     * consistent with `value.<name>.encodeToByteArray().size`; the
-     * codec trusts that contract (a runtime cross-check would
-     * allocate per row 16).
-     */
-    private fun appendEncodeLengthFromString(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthFromString,
-    ) {
-        body.addStatement(
-            "buffer.writeString(value.%L, %T.UTF8)",
-            field.name,
-            CHARSET_CN,
-        )
-    }
-
-    /**
-     * Emit decode for `@LengthFrom("siblingField")
-     * val: List<T>`. Bounds the buffer via `setLimit` to the
-     * sibling-derived byte count, loops reading elements via the
-     * element codec until the bounded position is reached, restores
-     * the outer limit. The `try`/`finally` guarantees limit
-     * restoration even if an element decode throws.
-     *
-     * Generated shape:
-     * ```
-     * <Int.MAX_VALUE guard for the sibling kind, if needed>
-     * val <name>Bytes = <sibling>.toInt()
-     * val <name>OuterLimit = buffer.limit()
-     * buffer.setLimit(buffer.position() + <name>Bytes)
-     * val <name> = mutableListOf<ElementType>()
-     * try {
-     *     while (buffer.position() < buffer.limit()) {
-     *         <name> += ElementCodec.decode(buffer, context)
-     *     }
-     * } finally {
-     *     buffer.setLimit(<name>OuterLimit)
-     * }
-     * ```
-     */
-    private fun appendDecodeLengthFromList(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthFromList,
-    ) {
-        if (field.source is LengthSource.Sibling) {
-            appendLengthFromIntMaxGuard(
-                body = body,
-                siblingAccessor = field.source.siblingName,
-                siblingKind = field.source.siblingKind,
-                ownerSimpleName = field.ownerSimpleName,
-                fieldName = field.name,
-            )
-        }
-        val bytesVar = "${field.name}Bytes"
-        val outerLimitVar = "${field.name}OuterLimit"
-        body.addStatement("val %L = %L", bytesVar, field.source.decodeAccessor())
-        body.addStatement("val %L = buffer.limit()", outerLimitVar)
-        body.addStatement("buffer.setLimit(buffer.position() + %L)", bytesVar)
-        body.addStatement("val %L = mutableListOf<%T>()", field.name, field.elementClassName)
-        body.beginControlFlow("try")
-        body.beginControlFlow("while (buffer.position() < buffer.limit())")
-        body.addStatement("%L += %T.decode(buffer, context)", field.name, field.elementCodecClassName)
-        body.endControlFlow()
-        body.nextControlFlow("finally")
-        body.addStatement("buffer.setLimit(%L)", outerLimitVar)
-        body.endControlFlow()
-    }
-
-    /**
-     * Emit encode for `@LengthFrom("siblingField")
-     * val: List<T>`. Iterates the list and writes each element via
-     * the element codec. The user is responsible for keeping
-     * `value.<sibling>` consistent with the sum of element wire
-     * sizes (same row-16 trust contract as the LengthFromString
-     * encode path).
-     */
-    private fun appendEncodeLengthFromList(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthFromList,
-    ) {
-        body.beginControlFlow("for (__elem in value.%L)", field.name)
-        body.addStatement("%T.encode(buffer, __elem, context)", field.elementCodecClassName)
-        body.endControlFlow()
-    }
-
-    /**
-     * (issue #151 part 1) — emit decode for
-     * `@LengthFrom("siblingField") val: T : @ProtocolMessage`. Bounds the
-     * buffer via `setLimit` to the sibling-derived end, delegates to
-     * `<TCodec>.decode(buffer, context)`, restores the outer limit in a
-     * `try`/`finally`. Same outer-limit-restore template as
-     * [appendDecodeLengthFromList].
-     *
-     * Generated shape:
-     * ```
-     * <Int.MAX_VALUE guard for the sibling kind, if needed>
-     * val <name>Bytes = <sibling>.toInt()
-     * val <name>OuterLimit = buffer.limit()
-     * buffer.setLimit(buffer.position() + <name>Bytes)
-     * val <name> = try {
-     *     <TCodec>.decode(buffer, context)
-     * } finally {
-     *     buffer.setLimit(<name>OuterLimit)
-     * }
-     * ```
-     */
-    private fun appendDecodeLengthFromMessage(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthFromMessage,
-    ) {
-        if (field.source is LengthSource.Sibling) {
-            appendLengthFromIntMaxGuard(
-                body = body,
-                siblingAccessor = field.source.siblingName,
-                siblingKind = field.source.siblingKind,
-                ownerSimpleName = field.ownerSimpleName,
-                fieldName = field.name,
-            )
-        }
-        val bytesVar = "${field.name}Bytes"
-        val outerLimitVar = "${field.name}OuterLimit"
-        body.addStatement("val %L = %L", bytesVar, field.source.decodeAccessor())
-        body.addStatement("val %L = buffer.limit()", outerLimitVar)
-        body.addStatement("buffer.setLimit(buffer.position() + %L)", bytesVar)
-        body.beginControlFlow("val %L = try", field.name)
-        body.addStatement("%T.decode(buffer, context)", field.codecType)
-        body.nextControlFlow("finally")
-        body.addStatement("buffer.setLimit(%L)", outerLimitVar)
-        body.endControlFlow()
-    }
-
-    /**
-     * Emit encode for `@LengthFrom("siblingField") val: T:
-     * @ProtocolMessage`. Single delegation to `<TCodec>.encode`. The
-     * sibling field has already been encoded by the prior field's emit
-     * step; the user is responsible for keeping `value.<sibling>`
-     * consistent with `<TCodec>.wireSize(value.<name>, context).bytes`.
-     */
-    private fun appendEncodeLengthFromMessage(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthFromMessage,
-    ) {
-        body.addStatement(
-            "%T.encode(buffer, value.%L, context)",
-            field.codecType,
-            field.name,
-        )
-    }
-
-    /**
-     * Emit decode for
-     * `@RemainingBytes @UseCodec(C::class) val: P`. Delegates to the
-     * user-supplied `C.decode(buffer, context)` against whatever
-     * `buffer.limit()` already says — same caller-bounds-buffer contract
-     * as the other `@RemainingBytes` shapes. The outer dispatcher (slice
-     * 10d for MQTT) sets the limit before calling this codec.
-     */
-    private fun appendDecodeRemainingBytesPayload(
-        body: CodeBlock.Builder,
-        field: FieldSpec.RemainingBytesPayload,
-    ) {
-        if (field.reservedTrailingBytes == 0) {
-            body.addStatement(
-                "val %L = %L.decode(buffer, context)",
-                field.name,
-                field.source.codecReceiver(),
-            )
-            return
-        }
-        // Non-terminal RemainingBytesPayload. Narrow the
-        // buffer's limit to leave the trailing FixedSize fields in the
-        // outer-limit region; restore the outer limit in a try/finally
-        // so the trailing field emits run against the original limit.
-        val outerLimitVar = "__${field.name}OuterLimit"
-        body.addStatement("val %L = buffer.limit()", outerLimitVar)
-        body.addStatement(
-            "buffer.setLimit(%L - %L)",
-            outerLimitVar,
-            field.reservedTrailingBytes,
-        )
-        body.beginControlFlow("val %L = try", field.name)
-        body.addStatement("%L.decode(buffer, context)", field.source.codecReceiver())
-        body.nextControlFlow("finally")
-        body.addStatement("buffer.setLimit(%L)", outerLimitVar)
-        body.endControlFlow()
-    }
-
-    /**
-     * Emit encode for
-     * `@RemainingBytes @UseCodec(C::class) val: P`. Delegates to the
-     * user-supplied `C.encode(buffer, value.<name>, context)`. No length
-     * carrier on the wire — the user codec writes its bytes against the
-     * buffer's current position and the trust contract (row 16) leaves
-     * total-byte-count consistency to the outer dispatcher.
-     */
-    private fun appendEncodeRemainingBytesPayload(
-        body: CodeBlock.Builder,
-        field: FieldSpec.RemainingBytesPayload,
-    ) {
-        body.addStatement(
-            "%L.encode(buffer, value.%L, context)",
-            field.source.codecReceiver(),
-            field.name,
-        )
-    }
-
-    /**
-     * `@RemainingBytes val: String` — decode reads UTF-8 bytes from the current
-     * position to `buffer.limit()`. The caller (or an outer dispatcher) is
-     * responsible for narrowing `buffer.limit()` to the bounded extent before
-     * invoking decode; same caller-bounds-buffer contract as
-     * [appendDecodeRemainingBytesPayload].
-     */
-    private fun appendDecodeRemainingBytesString(
-        body: CodeBlock.Builder,
-        field: FieldSpec.RemainingBytesString,
-    ) {
-        if (field.reservedTrailingBytes == 0) {
-            body.addStatement(
-                "val %L = buffer.readString(buffer.remaining(), %T.UTF8)",
-                field.name,
-                CHARSET_CN,
-            )
-            return
-        }
-        // Read the body byte count minus the reserved trailing
-        // FixedSize bytes; the trailing field emits run normally after.
-        body.addStatement(
-            "val %L = buffer.readString(buffer.remaining() - %L, %T.UTF8)",
-            field.name,
-            field.reservedTrailingBytes,
-            CHARSET_CN,
-        )
-    }
-
-    /**
-     * Encode counterpart for `@RemainingBytes val: String`. Writes the value's
-     * UTF-8 byte representation. The encoded byte count is reported via
-     * [appendBackPatchWireSize] (the parent message's wireSize collapses to
-     * BackPatch because the trailing string's byte count isn't known up front
-     * without re-encoding).
-     */
-    private fun appendEncodeRemainingBytesString(
-        body: CodeBlock.Builder,
-        field: FieldSpec.RemainingBytesString,
-    ) {
-        body.addStatement(
-            "buffer.writeString(value.%L, %T.UTF8)",
-            field.name,
-            CHARSET_CN,
-        )
-    }
-
-    /**
-     * Emit decode for `@RemainingBytes val: List<T>` where
-     * `T` is a `@ProtocolMessage data class`. Loops `while
-     * (buffer.position() < buffer.limit())` reading each element via
-     * the element's own codec. Caller-bounds-buffer contract: an outer
-     * dispatcher (e.g. MQTT's fixed-header remaining-length variable-
-     * length integer) sets `buffer.limit()` before delegating.
-     *
-     * Generated shape:
-     * ```
-     * val <name> = mutableListOf<ElementType>()
-     * while (buffer.position() < buffer.limit()) {
-     *     <name> += ElementCodec.decode(buffer, context)
-     * }
-     * ```
-     */
-    private fun appendDecodeRemainingBytesProtocolMessageList(
-        body: CodeBlock.Builder,
-        field: FieldSpec.RemainingBytesProtocolMessageList,
-    ) {
-        body.addStatement("val %L = mutableListOf<%T>()", field.name, field.elementClassName)
-        if (field.reservedTrailingBytes == 0) {
-            body.beginControlFlow("while (buffer.position() < buffer.limit())")
-        } else {
-            // Leave room for the trailing FixedSize fields.
-            body.beginControlFlow(
-                "while (buffer.position() < buffer.limit() - %L)",
-                field.reservedTrailingBytes,
-            )
-        }
-        body.addStatement("%L += %T.decode(buffer, context)", field.name, field.elementCodecClassName)
-        body.endControlFlow()
-    }
-
-    /**
-     * Emit encode for `@RemainingBytes val: List<T>` where
-     * `T` is a `@ProtocolMessage data class`. Iterates the list and
-     * writes each element via the element codec. The encoded byte
-     * count is implicit in the outer protocol's framing — same row 16
-     * trust contract as `LengthFromList`'s encode path.
-     */
-    private fun appendEncodeRemainingBytesProtocolMessageList(
-        body: CodeBlock.Builder,
-        field: FieldSpec.RemainingBytesProtocolMessageList,
-    ) {
-        body.beginControlFlow("for (__elem in value.%L)", field.name)
-        body.addStatement("%T.encode(buffer, __elem, context)", field.elementCodecClassName)
-        body.endControlFlow()
-    }
-
-    /**
-     * Emit decode for bare `@UseCodec val: <scalar>`.
-     * Delegates to the user-supplied codec object's `decode(buffer,
-     * context)`. When the codec implements [BoundingLengthCodec], the
-     * outer buffer limit is captured into `__<name>OuterLimit` BEFORE
-     * decode (so the surrounding try/finally restores the caller's
-     * outer limit even if the user codec or `applyBound` throws), and
-     * `applyBound(buffer, <name>)` runs after decode to narrow the
-     * limit for subsequent fields — driven by interface inspection on
-     * the codec target (the outer-limit-restore pattern).
-     */
-    private fun appendDecodeUseCodecScalar(
-        body: CodeBlock.Builder,
-        field: FieldSpec.UseCodecScalar,
-    ) {
-        if (field.isBounding) {
-            body.addStatement("val __%LOuterLimit = buffer.limit()", field.name)
-        }
-        body.addStatement("val %L = %T.decode(buffer, context)", field.name, field.codecType)
-        if (field.isBounding) {
-            body.addStatement("%T.applyBound(buffer, %L)", field.codecType, field.name)
-        }
-    }
-
-    /**
-     * Emit encode for bare `@UseCodec val: <scalar>`.
-     * Delegates to the user-supplied codec object's `encode(buffer,
-     * value.<name>, context)`. The user codec owns the wire shape;
-     * the framework neither validates nor measures the encoded width.
-     */
-    @Suppress("UNUSED_PARAMETER")
-    private fun appendEncodeUseCodecScalar(
-        body: CodeBlock.Builder,
-        field: FieldSpec.UseCodecScalar,
-        shape: CodecShape,
-    ) {
-        body.addStatement("%T.encode(buffer, value.%L, context)", field.codecType, field.name)
-    }
-
-    /**
-     * Emit decode/encode for bare `val: T:
-     * @ProtocolMessage`. Mirrors [appendEncodeUseCodecScalar] /
-     * [appendDecodeUseCodecScalar] minus the bounding-codec branch:
-     * the by-name-resolved codec is never a `BoundingLengthCodec` (those
-     * are user-supplied length codecs, never `@ProtocolMessage` body
-     * codecs).
-     */
-    private fun appendDecodeProtocolMessageScalar(
-        body: CodeBlock.Builder,
-        field: FieldSpec.ProtocolMessageScalar,
-    ) {
-        body.addStatement("val %L = %T.decode(buffer, context)", field.name, field.codecType)
-    }
-
-    private fun appendEncodeProtocolMessageScalar(
-        body: CodeBlock.Builder,
-        field: FieldSpec.ProtocolMessageScalar,
-    ) {
-        body.addStatement("%T.encode(buffer, value.%L, context)", field.codecType, field.name)
-    }
-
-    /**
-     * Emit decode for `@LengthPrefixed
-     * @UseCodec(C::class) val xs: List<E>`. The codec drives the prefix
-     * read and applies the resulting bound to `buffer.limit()`; the list
-     * is read element-by-element via E's codec inside the bounded region.
-     * Self-contained `try`/`finally` restores the outer limit, so
-     * subsequent fields run at the original limit.
-     *
-     * Generated shape:
-     * ```
-     * val __<name>OuterLimit = buffer.limit()
-     * val __<name>Length = <codecType>.decode(buffer, context)
-     * <codecType>.applyBound(buffer, __<name>Length)
-     * val <name> = mutableListOf<ElementType>()
-     * try {
-     *     while (buffer.position() < buffer.limit()) {
-     *         <name> += ElementCodec.decode(buffer, context)
-     *     }
-     * } finally {
-     *     buffer.setLimit(__<name>OuterLimit)
-     * }
-     * ```
-     */
-    private fun appendDecodeLengthPrefixedUseCodecList(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthPrefixedUseCodecList,
-    ) {
-        appendDecodeLengthPrefixedListBody(
-            body = body,
-            spec = field.spec,
-            listLocalName = field.name,
-            namespacePrefix = field.name,
-        )
-    }
-
-    /**
-     * Emit decode for `@LengthPrefixed
-     * @UseCodec(C::class) val: T : Payload`. Reads the fixed-width
-     * unsigned-int prefix, narrows `buffer.limit()` to position + length,
-     * delegates the body decode to `C.decode(buffer, context)`, and
-     * restores the outer limit in `try/finally`.
-     *
-     * Generated shape:
-     * ```
-     * val <name>Prefix = <prefix-decode>
-     * if (<name>Prefix > Int.MAX_VALUE.toUInt()) throw DecodeException(...)
-     * val <name>Length = <name>Prefix.toInt()
-     * val __<name>OuterLimit = buffer.limit()
-     * buffer.setLimit(buffer.position() + <name>Length)
-     * val <name> = try {
-     *     <PayloadCodec>.decode(buffer, context)
-     * } finally {
-     *     buffer.setLimit(__<name>OuterLimit)
-     * }
-     * ```
-     */
-    private fun appendDecodeLengthPrefixedUseCodecPayload(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthPrefixedUseCodecPayload,
-    ) {
-        val lengthVar =
-            appendLengthPrefixedStringPrefixDecode(
-                body = body,
-                name = field.name,
-                ownerSimpleName = field.ownerSimpleName,
-                prefixWidth = field.prefixWidth,
-                prefixWireOrder = field.prefixWireOrder,
-            )
-        val outerLimitVar = "__${field.name}OuterLimit"
-        body.addStatement("val %L = buffer.limit()", outerLimitVar)
-        body.addStatement("buffer.setLimit(buffer.position() + %L)", lengthVar)
-        body.beginControlFlow("val %L = try", field.name)
-        body.addStatement("%T.decode(buffer, context)", field.payloadCodecType)
-        body.nextControlFlow("finally")
-        body.addStatement("buffer.setLimit(%L)", outerLimitVar)
-        body.endControlFlow()
-    }
-
-    /**
-     * Shared decode body for the VBI-prefixed
-     * list shape. Emitted by both `FieldSpec.LengthPrefixedUseCodecList`
-     *  and the conditional-inner branch in
-     * `appendDecodeConditional`. Five-step sequence:
-     * capture outer limit → codec.decode VBI prefix → applyBound →
-     * mutableListOf → try-while-finally restore outer limit.
-     *
-     * `listLocalName` is the variable that holds the decoded list. The
-     * non-conditional path uses the field's own name (`<field>`); the
-     * conditional path uses `<field>Value` because `<field>` is a
-     * nullable-typed local that `appendDecodeConditional` sets via the
-     * `if (predicate) { ... <listLocal> } else null` construction.
-     *
-     * `namespacePrefix` keys the local-variable names (`__<prefix>
-     * OuterLimit`, `__<prefix>Length`). Field path passes the field
-     * name; conditional path also passes the field name (so encode/
-     * decode share scratch local names within the same conditional
-     * slot).
-     */
-    private fun appendDecodeLengthPrefixedListBody(
-        body: CodeBlock.Builder,
-        spec: LengthPrefixedListSpec,
-        listLocalName: String,
-        namespacePrefix: String,
-    ) {
-        val outerLimitVar = "__${namespacePrefix}OuterLimit"
-        val lengthVar = "__${namespacePrefix}Length"
-        body.addStatement("val %L = buffer.limit()", outerLimitVar)
-        body.addStatement("val %L = %T.decode(buffer, context)", lengthVar, spec.codecType)
-        body.addStatement("%T.applyBound(buffer, %L)", spec.codecType, lengthVar)
-        body.addStatement("val %L = mutableListOf<%T>()", listLocalName, spec.elementClassName)
-        body.beginControlFlow("try")
-        body.beginControlFlow("while (buffer.position() < buffer.limit())")
-        body.addStatement("%L += %T.decode(buffer, context)", listLocalName, spec.elementCodecClassName)
-        body.endControlFlow()
-        body.nextControlFlow("finally")
-        body.addStatement("buffer.setLimit(%L)", outerLimitVar)
-        body.endControlFlow()
-    }
-
-    /**
-     * Emit encode for `@LengthPrefixed
-     * @UseCodec(C::class) val xs: List<E>`. Pre-measures the body byte
-     * count via the element codec's `wireSize` (cast to `Exact`), writes
-     * the prefix via the user codec's `encode`, then iterates and encodes
-     * elements. BackPatch element codecs throw `ClassCastException` —
-     * same fixture-design contract as `RemainingBytesProtocolMessageList`
-     * and `LengthPrefixedMessage`.
-     *
-     * Generated shape:
-     * ```
-     * val __<name>BodyBytes = value.<name>.sumOf {
-     *     (ElementCodec.wireSize(it, context) as WireSize.Exact).bytes
-     * }
-     * <codecType>.encode(buffer, __<name>BodyBytes.toUInt(), context)
-     * for (__elem in value.<name>) {
-     *     ElementCodec.encode(buffer, __elem, context)
-     * }
-     * ```
-     */
-    private fun appendEncodeLengthPrefixedUseCodecList(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthPrefixedUseCodecList,
-    ) {
-        appendEncodeLengthPrefixedListBody(
-            body = body,
-            spec = field.spec,
-            accessor = "value.${field.name}",
-            namespacePrefix = field.name,
-        )
-    }
-
-    /**
-     * Emit encode for `@LengthPrefixed
-     * @UseCodec(C::class) val: T : Payload`. BackPatch shape mirroring
-     * [appendLengthPrefixedStringEncode]: reserve prefix slot, run
-     * `C.encode(buffer, value.<name>, context)` against the accumulating
-     * buffer, measure the body byte count from the position delta,
-     * patch the prefix in place, restore position past the body.
-     */
-    private fun appendEncodeLengthPrefixedUseCodecPayload(
-        body: CodeBlock.Builder,
-        field: FieldSpec.LengthPrefixedUseCodecPayload,
-    ) {
-        val sizePosVar = "${field.name}SizePosition"
-        val bodyStartVar = "${field.name}BodyStart"
-        val endPosVar = "${field.name}EndPosition"
-        val byteCountVar = "${field.name}ByteCount"
-        body.addStatement("val %L = buffer.position()", sizePosVar)
-        body.addStatement("buffer.position(%L + %L)", sizePosVar, field.prefixWidth)
-        body.addStatement("val %L = buffer.position()", bodyStartVar)
-        body.addStatement(
-            "%T.encode(buffer, value.%L, context)",
-            field.payloadCodecType,
-            field.name,
-        )
-        body.addStatement("val %L = buffer.position()", endPosVar)
-        body.addStatement("val %L = %L - %L", byteCountVar, endPosVar, bodyStartVar)
-        if (field.prefixWidth < 4) {
-            val maxValue = (1L shl (field.prefixWidth * 8)) - 1
-            val widthName =
-                when (field.prefixWidth) {
-                    1 -> "Byte"
-                    2 -> "Short"
-                    else -> error("unreachable: prefixWidth must be 1, 2, or 4")
-                }
-            body.beginControlFlow("if (%L > %L)", byteCountVar, maxValue)
-            body.addStatement(
-                "throw %T(fieldPath = %S, reason = %P)",
-                ENCODE_EXCEPTION_CN,
-                "${field.ownerSimpleName}.${field.name}",
-                "encoded payload byte length \${$byteCountVar} exceeds " +
-                    "@LengthPrefixed(LengthPrefix.$widthName) max $maxValue",
-            )
-            body.endControlFlow()
-        }
-        body.addStatement("buffer.position(%L)", sizePosVar)
-        val prefixVar = "${field.name}Prefix"
-        body.addStatement("val %L = %L.toUInt()", prefixVar, byteCountVar)
-        appendBufferPrefixEncode(body, prefixVar, field.prefixWidth, field.prefixWireOrder)
-        body.addStatement("buffer.position(%L)", endPosVar)
-    }
-
-    /**
-     * Shared encode body for the VBI-prefixed
-     * list shape. Emitted by both `FieldSpec.LengthPrefixedUseCodecList`
-     *  and `appendEncodeConditional`'s
-     * `LengthPrefixedUseCodecList` branch.
-     *
-     * `accessor` is the read-side expression for the list — `value.
-     * <name>` for the non-conditional path; the smart-cast non-null
-     * local (`<name>Value`) for the conditional path. `namespacePrefix`
-     * keys the scratch / body-bytes locals.
-     *
-     * Two encode paths, gated by `spec.elementIsBackPatch`:
-     *
-     * **Sealed elements** — variants commonly carry BackPatch-wireSize
-     * fields (`@LengthPrefixed val: String`, `@When` trailers), so the
-     * pre-measure `as WireSize.Exact` cast doesn't apply. Encode each
-     * element into a scratch buffer first to capture the actual byte
-     * count, then write the VBI prefix and bulk-copy:
-     * ```
-     * BufferFactory.Default.allocate(64, buffer.byteOrder).use { __<n>Scratch ->
-     *     for (__elem in <accessor>) {
-     *         ElementCodec.encode(__<n>Scratch, __elem, context)
-     *     }
-     *     val __<n>BodyBytes = __<n>Scratch.position()
-     *     <codecType>.encode(buffer, __<n>BodyBytes.toUInt(), context)
-     *     __<n>Scratch.resetForRead()
-     *     buffer.write(__<n>Scratch)
-     * }
-     * ```
-     * 64-byte starting allocation is a heuristic — `BufferFactory` grows
-     * on demand for buffers that exceed it. Tunable per-field if a
-     * measurable hot path emerges.
-     *
-     * **Data-class elements** — pre-measure body bytes via the element
-     * codec's `wireSize as Exact`, write VBI prefix, iterate. BackPatch
-     * elements throw `ClassCastException` — same fixture-design contract
-     * as `RemainingBytesProtocolMessageList` and `LengthPrefixedMessage`.
-     * Audit 2b notes the latent risk: a data-class element with a
-     * `@LengthPrefixed val: String` field has BackPatch wireSize and
-     * would CCE; no current fixture trips it because all sealed-parent
-     * cases are routed through the scratch path.
-     */
-    private fun appendEncodeLengthPrefixedListBody(
-        body: CodeBlock.Builder,
-        spec: LengthPrefixedListSpec,
-        accessor: String,
-        namespacePrefix: String,
-    ) {
-        if (spec.elementIsBackPatch) {
-            val scratchVar = "__${namespacePrefix}Scratch"
-            val bodyBytesVar = "__${namespacePrefix}BodyBytes"
-            body.beginControlFlow(
-                "%T.%M.allocate(64, buffer.byteOrder).%M { %L ->",
-                BUFFER_FACTORY_CN,
-                BUFFER_FACTORY_DEFAULT_MN,
-                BUFFER_USE_MN,
-                scratchVar,
-            )
-            body.beginControlFlow("for (__elem in %L)", accessor)
-            body.addStatement(
-                "%T.encode(%L, __elem, context)",
-                spec.elementCodecClassName,
-                scratchVar,
-            )
-            body.endControlFlow()
-            body.addStatement("val %L = %L.position()", bodyBytesVar, scratchVar)
-            body.addStatement(
-                "%T.encode(buffer, %L.toUInt(), context)",
-                spec.codecType,
-                bodyBytesVar,
-            )
-            body.addStatement("%L.resetForRead()", scratchVar)
-            body.addStatement("buffer.write(%L)", scratchVar)
-            body.endControlFlow()
-        } else {
-            val bodyBytesVar = "__${namespacePrefix}BodyBytes"
-            body.addStatement(
-                "val %L = %L.sumOf { (%T.wireSize(it, context) as %T.Exact).bytes }",
-                bodyBytesVar,
-                accessor,
-                spec.elementCodecClassName,
-                WIRE_SIZE_CN,
-            )
-            body.addStatement(
-                "%T.encode(buffer, %L.toUInt(), context)",
-                spec.codecType,
-                bodyBytesVar,
-            )
-            body.beginControlFlow("for (__elem in %L)", accessor)
-            body.addStatement("%T.encode(buffer, __elem, context)", spec.elementCodecClassName)
-            body.endControlFlow()
-        }
-    }
-
-    private fun appendBufferPrefixEncode(
-        body: CodeBlock.Builder,
-        prefixVar: String,
-        prefixWidth: Int,
-        wireOrder: Endianness,
-    ) {
-        if (prefixWidth == 1) {
-            body.addStatement("buffer.writeUByte((%L and 0xFFu).toUByte())", prefixVar)
-            return
-        }
-        val bigEndian =
-            when (wireOrder) {
-                Endianness.Big -> true
-                Endianness.Little -> false
-                Endianness.Default -> true
-            }
-        for (i in 0 until prefixWidth) {
-            val shiftBits = if (bigEndian) (prefixWidth - 1 - i) * 8 else i * 8
-            val expr =
-                if (shiftBits == 0) {
-                    "$prefixVar and 0xFFu"
-                } else {
-                    "($prefixVar shr $shiftBits) and 0xFFu"
-                }
-            body.addStatement("buffer.writeUByte((%L).toUByte())", expr)
-        }
     }
 }
